Remove unnecessary duplicat files.
The types were copy pasted and there was an excess of files.
