+++ homesensorhub/sensors/types/BoschBME/BoschBME280/altitude_BoschBME280.py
@@ -1,5 +1,5 @@
 """Module which implements the Bosch280BME altitude sensor module set up."""
-from sensors.types.altitude import Altitude
+from sensors.types.sensor_type import Altitude
 
 
 class AltitudeBoschBME280(Altitude):
