+++ homesensorhub/sensors/types/BoschBME/BoschBME280/humidity_BoschBME280.py
@@ -1,5 +1,5 @@
 """Module which implements the Bosch280BME humidity sensor module set up."""
-from sensors.types.humidity import Humidity
+from sensors.types.sensor_type import Humidity
 
 
 class HumidityBoschBME280(Humidity):
