+++ homesensorhub/sensors/types/BoschBME/BoschBME280/temperature_BoschBME280.py
@@ -1,5 +1,5 @@
 """Module which implements the Bosch280BME temperature sensor module logic."""
-from sensors.types.temperature import Temperature
+from sensors.types.sensor_type import Temperature
 
 
 class TemperatureBoschBME280(Temperature):
